build: use correct layout for a Python namespace package
PDM does not properly handle the combination of using a src/ directory
with a namespace package and fails to build the package properly. The
filesystem layout has been adjusted to reflect this:

  `src/feeph/i2c/` -> `feeph/i2c/`

With this modified layout the build will work properly and mypy is able to
find all relevant files.