feat: redesign the library and document its philosophy
- replaces `SimulatedI2cBus()` with `EmulatedI2C()`
- 'HardwareI2cBus()` has been removed

The library has been redesigned and it's no longer necessary to hide
`busio.I2C()` behind another abstraction layer. As part of this change and
to indicate the relationship to `busio.I2C()` the class has been renamed
from `SimulatedI2cBus()` to `EmulatedI2C()` and inherits from `busio.I2C()`
instead of `I2cBusInterface`.

Additionally 4 functions have been added that can be used with both `I2C()`
and `EmulatedI2cBus()`. The functions make it more convenient and less
confusing to work with busio's I²C bus interface.

Technically this is a breaking change but we're still in 0.x (where major
changes can be expected) and there are no 3rd party users of this library
since it hasn't been released on a public repository yet. We can break
compatibility and nobody will care.